+++ Assets/Scenes/AutoSave/Level_Tutorial Dec 11-30am.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4eae1805e7288fb4c962d65f125e6839
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
