--- Assets/Scripts/Tutorial/DialogueData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9af793e73521b5245b9fc354f78924f7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
