--- Assets/Scripts/Tutorial/DialogueManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 516daaa205b4fbe408df621bc8553dbd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
