--- Assets/Scripts/Tutorial/DialogueTrigger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3974ae4e0e5b6bb48ba34d86a3922b4a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
