--- Assets/Scripts/Tutorial/TutDia_01.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7b80d553c2050114e8eab4f64e8fc299
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
