--- Assets/Scripts/Tutorial/TutDia_03.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 13813d00d6d5e6c4791196ac6a1c5746
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
