--- Assets/Scripts/Tutorial/TutDia_04.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 158cf45a1fd91914081679f6d135169a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
