--- Assets/Scripts/Tutorial/TutDia_06.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2103a93fd5b4cf34984468b816ea7b84
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
