--- Assets/Scripts/Tutorial/TutDia_08.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 68271b606a8ef8747b1ea73ed33cb9cb
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
