+++ Assets/Scripts/Tutorial/Tutorial Dialogue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aadf9d6fa3d13944e9b50647d69a2d11
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
