+++ Assets/Scripts/Tutorial/Tutorial Dialogue/DialogueData.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+[CreateAssetMenu(fileName = "NewDialogue", menuName = "ScriptableObjects/Tutorial Dialogue")]
+public class DialogueData : ScriptableObject
+{
+    public List<string> dialogueIDs;
+}
