+++ Assets/Scripts/Tutorial/Tutorial Dialogue/TutDia_05.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cdac5b2772746d445b808c919f1936ce
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
