+++ Assets/Scripts/Tutorial/Tutorial Dialogue/TutDia_06.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2103a93fd5b4cf34984468b816ea7b84
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
