+++ Assets/Scripts/Tutorial/Tutorial Dialogue/TutDia_07.asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9af793e73521b5245b9fc354f78924f7, type: 3}
+  m_Name: TutDia_07
+  m_EditorClassIdentifier: 
+  dialogueIDs:
+  - TUTDialogue07
