+++ Assets/Scripts/Tutorial/Tutorial Dialogue/TutDia_07.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fd390513f4175654498ead706bea48b3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
