+++ Assets/Scripts/Tutorial/Tutorial Dialogue/TutDia_08.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 68271b606a8ef8747b1ea73ed33cb9cb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
