+++ Assets/Scripts/Tutorial/Tutorial Dialogue/TutDia_09.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d596622bc4b9b94d81b904511a581e0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
